refactor code browseBase schema and add rowCollapse schema prop
