+++ lib/schemas/common/browseBase/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const browseBase = require('./browseBase');
+
+module.exports = browseBase;
